Pack proving key from arkworks
