+++ groth16/src/lib.rs
@@ -3,6 +3,8 @@ use ark_poly::{domain::EvaluationDomain, Radix2EvaluationDomain};
 
 pub mod ext_wit;
 pub mod pre_processing;
+pub mod prover;
+pub mod proving_key;
 pub mod qap;
 
 #[derive(Debug, Clone, PartialEq)]
